requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic-settings==2.11.0
sentence-transformers==5.1.1
faiss-cpu==1.12.0
httpx==0.28.1
tree-sitter-language-pack==1.2.1
keyring==25.6.0
